Remove WebUtil.load_scripts
The only remaining user of WebUtil.load_scripts was for loading
localisation.  The ES6 modules spec has support for dynamic loading
of modules, so we use that instead (this required removing the
window-level "Language" object, but that should never have been there in
the first place).

Note that none of the Babel plugins besides the SystemJS one support
translating dynamic import statements, so if packaging the full app
as CommonJS, one should load the language module separately.